+++ erp.backend/src/test/java/org/roko/erp/backend/controllers/BankAccountControllerTest.java
@@ -32,8 +32,8 @@ public class BankAccountControllerTest {
 
     private static final int TEST_PAGE = 123;
 
-    private static final int TEST_COUNT = 222;
-    private static final int TEST_LEDGER_ENTRY_COUNT = 333;
+    private static final long TEST_COUNT = 222;
+    private static final long TEST_LEDGER_ENTRY_COUNT = 333;
 
     @Mock
     private BankAccount bankAccountMock;
