+++ erp.backend/src/test/java/org/roko/erp/backend/controllers/PostedPurchaseOrderControllerTest.java
@@ -21,7 +21,7 @@ import org.roko.erp.dto.list.PostedPurchaseDocumentList;
 
 public class PostedPurchaseOrderControllerTest {
 
-    private static final int TEST_COUNT = 222;
+    private static final long TEST_COUNT = 222;
 
     private static final String TEST_CODE = "test-code";
